@@ -19,13 +19,13 @@
     "url": "https://github.com/sitecore/jss/issues"
   },
   "devDependencies": {
-    "@angular/common": "^8.2.8",
-    "@angular/compiler": "^8.2.8",
-    "@angular/compiler-cli": "^8.2.8",
-    "@angular/core": "^8.2.8",
-    "@angular/platform-browser": "^8.2.8",
-    "@angular/platform-browser-dynamic": "^8.2.8",
-    "@angular/router": "^8.2.8",
+    "@angular/common": "~10.0.11",
+    "@angular/compiler": "~10.0.11",
+    "@angular/compiler-cli": "~10.0.11",
+    "@angular/core": "~10.0.11",
+    "@angular/platform-browser": "~10.0.11",
+    "@angular/platform-browser-dynamic": "~10.0.11",
+    "@angular/router": "~10.0.11",
     "@types/jasmine": "^3.4.1",
     "@types/node": "^12.7.8",
     "codelyzer": "^4.5.0",
@@ -36,18 +36,18 @@
     "karma-firefox-launcher": "^1.2.0",
     "karma-jasmine": "^2.0.1",
     "karma-typescript": "^5.0.3",
-    "ng-packagr": "^5.5.1",
+    "ng-packagr": "^10.0.4",
     "reflect-metadata": "^0.1.13",
     "rxjs": "~6.5.3",
     "tsickle": "^0.37.0",
-    "typescript": "~3.5.0",
+    "typescript": "~3.9.7",
     "url-parse": "^1.4.7",
-    "zone.js": "~0.10.2"
+    "zone.js": "~0.10.3"
   },
   "peerDependencies": {
-    "@angular/common": "^8.0.0",
-    "@angular/core": "^8.0.0",
-    "@angular/router": "^8.0.0",
+    "@angular/common": "~10.0.11",
+    "@angular/core": "~10.0.11",
+    "@angular/router": "~10.0.11",
     "rxjs": "^6.5.3"
   },
   "dependencies": {
