@@ -53,25 +53,23 @@
   "homepage": "https://jss.sitecore.net",
   "license": "Apache-2.0",
   "dependencies": {
-    "@angular/animations": "~8.2.8",
-    "@angular/common": "~8.2.8",
-    "@angular/compiler": "~8.2.8",
-    "@angular/core": "~8.2.8",
-    "@angular/forms": "~8.2.8",
-    "@angular/http": "~7.2.15",
-    "@angular/platform-browser": "~8.2.8",
-    "@angular/platform-browser-dynamic": "~8.2.8",
-    "@angular/platform-server": "~8.2.8",
-    "@angular/router": "~8.2.8",
-    "@nguniversal/module-map-ngfactory-loader": "~8.1.1",
-    "@ngx-meta/core": "~7.0.0",
+    "@angular/animations": "~10.0.11",
+    "@angular/common": "~10.0.11",
+    "@angular/compiler": "~10.0.11",
+    "@angular/core": "~10.0.11",
+    "@angular/forms": "~10.0.11",
+    "@angular/platform-browser": "~10.0.11",
+    "@angular/platform-browser-dynamic": "~10.0.11",
+    "@angular/platform-server": "~10.0.11",
+    "@angular/router": "~10.0.11",
+    "@ngx-meta/core": "~9.0.0",
     "@ngx-translate/core": "~11.0.1",
     "@ngx-translate/http-loader": "~4.0.0",
     "@sitecore-jss/sitecore-jss": "^14.0.1",
     "@sitecore-jss/sitecore-jss-angular": "^14.0.1",
     "@sitecore-jss/sitecore-jss-tracking": "^14.0.1",
-    "apollo-angular": "~1.7.0",
-    "apollo-angular-link-http-batch": "~1.7.0",
+    "apollo-angular": "~1.10.0",
+    "apollo-angular-link-http-batch": "~1.9.0",
     "apollo-angular-link-persisted": "~1.1.0",
     "apollo-cache-inmemory": "~1.6.3",
     "apollo-client": "~2.6.4",
@@ -118,8 +116,8 @@
     "protractor": "~5.4.2",
     "ts-loader": "~6.2.0",
     "ts-node": "~8.4.1",
-    "tslint": "~5.20.0",
-    "typescript": "~3.5.0",
+    "tslint": "~6.1.3",
+    "typescript": "~3.9.7",
     "webpack-cli": "~3.3.9",
     "yaml-lint": "^1.2.4"
   }
