@@ -4,7 +4,7 @@ import { CommonFieldTypes, SitecoreIcon, Manifest } from '@sitecore-jss/sitecore
  * Adds the Styleguide-CustomRouteType component to the disconnected manifest.
  * This function is invoked by convention (*.sitecore.ts) when `jss manifest` is run.
  */
-export default function(manifest: Manifest) {
+export default function StyleguideCustomRouteType(manifest: Manifest) {
   // This registers the custom route type with the manifest.
   // Like a component, a route type contains data fields -
   // but unlike a component, the fields are shared at the route level,
