@@ -5,7 +5,7 @@ const packageJson = require('../../../package.json');
  * Adds the Styleguide-FieldUsage-ContentList component to the disconnected manifest.
  * This function is invoked by convention (*.sitecore.ts) when `jss manifest` is run.
  */
-export default function(manifest: Manifest) {
+export default function StyleguideFieldUsageContentList(manifest: Manifest) {
   manifest.addComponent({
     name: 'StyleguideFieldUsageContentList',
     icon: SitecoreIcon.ListStyle_numbered,
