@@ -4,7 +4,7 @@ import { SitecoreIcon, Manifest } from '@sitecore-jss/sitecore-jss-manifest';
  * Adds the StyleguideRouteFields component to the disconnected manifest.
  * This function is invoked by convention (*.sitecore.ts) when `jss manifest` is run.
  */
-export default function(manifest: Manifest) {
+export default function StyleguideRouteFields(manifest: Manifest) {
   manifest.addComponent({
     name: 'StyleguideRouteFields',
     icon: SitecoreIcon.TextField,
