@@ -4,7 +4,7 @@ import { SitecoreIcon, Manifest } from '@sitecore-jss/sitecore-jss-manifest';
  * Adds the Styleguide-Tracking component to the disconnected manifest.
  * This function is invoked by convention (*.sitecore.js) when 'jss manifest' is run.
  */
-export default function(manifest: Manifest) {
+export default function StyleguideTracking(manifest: Manifest) {
   manifest.addComponent({
     name: 'StyleguideTracking',
     icon: SitecoreIcon.Compass,
