@@ -3,4 +3,6 @@
 
 console.log('Enabling TypeScript transpilation for the manifest...');
 
-require('ts-node/register');
+require('ts-node').register({
+	project: 'src/tsconfig.webpack-server.json'
+});
