@@ -18,7 +18,7 @@ export default function addContentToManifest(manifest: Manifest): Promise<any> {
 
   return mergeFs(startPath)
     .then((result) => {
-      const items = convertToItems(
+      const items = convertToContentItems(
         result,
         resolvePath(startPath),
         rootItemName,
@@ -36,8 +36,8 @@ export default function addContentToManifest(manifest: Manifest): Promise<any> {
 /**
  * Maps filesystem content data into manifest content item data.
  */
-function convertToItems(data: MergeFsResult, basePath: string, rootItemName: string, language: string): ItemDefinition {
-  const itemPath = convertPhsyicalPathToItemRelativePath(data.path, basePath);
+function convertToContentItems(data: MergeFsResult, basePath: string, rootItemName: string, language: string): ItemDefinition {
+  const itemPath = convertPhysicalContentPathToItemRelativePath(data.path, basePath);
   const name = itemPath.substr(itemPath.lastIndexOf('/') + 1);
 
   let result;
@@ -74,7 +74,7 @@ function convertToItems(data: MergeFsResult, basePath: string, rootItemName: str
   // recursively process child paths
   if (data.folders.length > 0) {
     result.children = data.folders
-      .map((folder) => convertToItems(folder, basePath, rootItemName, language))
+      .map((folder) => convertToContentItems(folder, basePath, rootItemName, language))
       .filter((item) => item); // remove null results
   }
 
@@ -85,7 +85,7 @@ function convertToItems(data: MergeFsResult, basePath: string, rootItemName: str
  * Converts a physical filesystem path into a relative Sitecore item path.
  * i.e. if physicalPath = /var/log and basePath = /var, this returns /log.
  */
-function convertPhsyicalPathToItemRelativePath(physicalPath: string, basePath: string) {
+function convertPhysicalContentPathToItemRelativePath(physicalPath: string, basePath: string) {
   const targetPathSeparator = '/';
 
   // normalize path separators to /
